Moved more unit tests to align with the renderer they really test, rather than posing as CukedoctorConverterImpl tests
