+++ cukedoctor-converter/src/test/java/com/github/cukedoctor/renderer/CukedoctorFeatureRendererTest.java
@@ -2,10 +2,12 @@ package com.github.cukedoctor.renderer;
 
 import com.github.cukedoctor.api.CukedoctorDocumentBuilder;
 import com.github.cukedoctor.api.DocumentAttributes;
-import com.github.cukedoctor.api.model.Feature;
-import com.github.cukedoctor.api.model.Tag;
+import com.github.cukedoctor.api.model.*;
+import com.github.cukedoctor.builder.CukedoctorDocumentBuilderImpl;
 import com.github.cukedoctor.parser.FeatureParser;
+import com.github.cukedoctor.util.Expectations;
 import com.github.cukedoctor.util.builder.FeatureBuilder;
+import com.github.cukedoctor.util.builder.ScenarioBuilder;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
@@ -13,8 +15,12 @@ import org.junit.runners.JUnit4;
 import java.util.ArrayList;
 import java.util.List;
 
+import static com.github.cukedoctor.renderer.Fixtures.featureWithSourceDocStringInStep;
 import static com.github.cukedoctor.util.Constants.newLine;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Matchers.*;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.spy;
 
 @RunWith(JUnit4.class)
 public class CukedoctorFeatureRendererTest {
@@ -198,5 +204,252 @@ public class CukedoctorFeatureRendererTest {
                 "==========" + newLine() +
                 "" + newLine() +
                 "").replaceAll("\r", ""));
+    }
+
+    @Test
+    public void shouldRenderFeatureScenarios() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithTwoScenarios();
+
+        CukedoctorScenarioRenderer scenarioRenderer = new CukedoctorScenarioRenderer();
+        scenarioRenderer = spy(scenarioRenderer);
+        doReturn("").when(scenarioRenderer).renderScenarioSteps(anyListOf(Step.class), any(Scenario.class), any(Feature.class));
+        doReturn("").when(scenarioRenderer).renderScenarioExamples(any(Scenario.class));
+        doReturn("").when(scenarioRenderer).renderScenarioTags(any(Scenario.class), eq(feature));
+
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle().nestTitle());
+        featureRenderer = spy(featureRenderer);
+        featureRenderer.scenarioRenderer = scenarioRenderer;
+
+        String resultDoc = featureRenderer.renderFeatureScenarios(feature, featureRenderer.docBuilder.createNestedBuilder());
+        assertThat(resultDoc).isEqualTo("==== Scenario: scenario 1" + newLine() +
+                "description" + newLine() + newLine() +
+                "==== Scenario: scenario 2" + newLine() +
+                "description 2" + newLine() + newLine());
+    }
+
+    @Test
+    public void shouldRenderFeatureScenariosWithTagsInScenarios() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithTwoScenarios();
+        for (Scenario scenario : feature.getScenarios()) {
+            ScenarioBuilder.instance(scenario).tag(new Tag("@Tag1")).tag(new Tag("@tag2"));
+        }
+        CukedoctorScenarioRenderer scenarioRenderer = new CukedoctorScenarioRenderer();
+        scenarioRenderer = spy(scenarioRenderer);
+        doReturn("").when(scenarioRenderer).renderScenarioSteps(anyListOf(Step.class), any(Scenario.class), any(Feature.class));
+        doReturn("").when(scenarioRenderer).renderScenarioExamples(any(Scenario.class));
+
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle().nestTitle());
+        featureRenderer = spy(featureRenderer);
+        featureRenderer.scenarioRenderer = scenarioRenderer;
+
+        String resultDoc = featureRenderer.renderFeatureScenarios(feature, featureRenderer.docBuilder.createNestedBuilder());
+        assertThat(resultDoc).isEqualTo("==== Scenario: scenario 1" + newLine() +
+                "[small]#tags: @tag2,@Tag1#" + newLine() +
+                "" + newLine() +
+                "description" + newLine() + newLine() +
+                "==== Scenario: scenario 2" + newLine() +
+                "[small]#tags: @tag2,@Tag1#" + newLine() +
+                "" + newLine() +
+                "description 2" + newLine() + newLine());
+    }
+
+    @Test
+    public void shouldRenderFeatureScenariosWithTagsInFeaturesAndScenarios() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithTwoScenarios();
+        feature.getTags().add(new Tag("@FeatureTag"));
+        for (Scenario scenario : feature.getScenarios()) {
+            ScenarioBuilder.instance(scenario).tag(new Tag("@Tag1")).tag(new Tag("@tag2"));
+        }
+
+        CukedoctorScenarioRenderer scenarioRenderer = new CukedoctorScenarioRenderer();
+        scenarioRenderer = spy(scenarioRenderer);
+        doReturn("").when(scenarioRenderer).renderScenarioSteps(anyListOf(Step.class), any(Scenario.class), any(Feature.class));
+        doReturn("").when(scenarioRenderer).renderScenarioExamples(any(Scenario.class));
+
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle().nestTitle());
+        featureRenderer = spy(featureRenderer);
+        featureRenderer.scenarioRenderer = scenarioRenderer;
+
+        String resultDoc = featureRenderer.renderFeatureScenarios(feature, featureRenderer.docBuilder.createNestedBuilder());
+        assertThat(resultDoc).isEqualTo("==== Scenario: scenario 1" + newLine() +
+                "[small]#tags: @tag2,@FeatureTag,@Tag1#" + newLine() +
+                "" + newLine() +
+                "description" + newLine() + newLine() +
+                "==== Scenario: scenario 2" + newLine() +
+                "[small]#tags: @tag2,@FeatureTag,@Tag1#" + newLine() +
+                "" + newLine() +
+                "description 2" + newLine() + newLine());
+    }
+
+    @Test
+    public void shouldRenderFeatureScenariosOmittingTagsIfOrderFeatureTagIsTheOnlyTag() {
+        // cucumber-jvm-5.6.0 cascades feature tags down to scenarios. See "ordering.feature" and "order-tags.json" for an example.
+        final Feature feature = FeatureBuilder.instance().aFeatureWithTwoScenarios();
+        final String orderTagName = "@order-42";
+        feature.getTags().add(new Tag(orderTagName));
+        for (Scenario scenario : feature.getScenarios()) {
+            ScenarioBuilder.instance(scenario).tag(new Tag(orderTagName));
+        }
+
+        CukedoctorScenarioRenderer scenarioRenderer = new CukedoctorScenarioRenderer();
+        scenarioRenderer = spy(scenarioRenderer);
+        doReturn("").when(scenarioRenderer).renderScenarioSteps(anyListOf(Step.class), any(Scenario.class), any(Feature.class));
+        doReturn("").when(scenarioRenderer).renderScenarioExamples(any(Scenario.class));
+
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle().nestTitle());
+        featureRenderer = spy(featureRenderer);
+        featureRenderer.scenarioRenderer = scenarioRenderer;
+
+        String resultDoc = featureRenderer.renderFeatureScenarios(feature, featureRenderer.docBuilder.createNestedBuilder());
+        assertThat(resultDoc).isEqualTo("==== Scenario: scenario 1" + newLine() +
+                "description" + newLine() + newLine() +
+                "==== Scenario: scenario 2" + newLine() +
+                "description 2" + newLine() + newLine());
+    }
+
+    @Test
+    public void shouldNotRenderScenarioWithSkipDocsTag() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithOneScenarioWithOnePassingStep();
+        Scenario scenarioToSkip = ScenarioBuilder.instance().name("scenario to skip").tag(new Tag("@skipDocs")).
+                type(Type.scenario).build();
+        ;
+        feature.getScenarios().add(scenarioToSkip);
+        List<Feature> features = new ArrayList<>();
+        features.add(feature);
+
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle().nestTitle());
+        String resultDoc = featureRenderer.renderFeatureScenarios(feature, featureRenderer.docBuilder.createNestedBuilder());
+
+        assertThat(resultDoc).
+                doesNotContain("scenario to skip").
+                isEqualTo("==== Scenario: scenario" + newLine() +
+                        "description" + newLine() + newLine() +
+                        "==========" + newLine() +
+                        "Given::" + newLine() +
+                        "passing step icon:thumbs-up[role=\"green\",title=\"Passed\"] [small right]#(000ms)#" + newLine() +
+                        "==========" + newLine() +
+                        "" + newLine());
+    }
+
+    @Test
+    public void shouldRenderFeatureScenariosWithMultipleSteps() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithMultipleScenariosAndSteps();
+        List<Feature> features = new ArrayList<>();
+        features.add(feature);
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle().nestTitle());
+        String resultDoc = featureRenderer.renderFeatureScenarios(feature, featureRenderer.docBuilder);
+        assertThat(resultDoc.replaceAll("\r", "")).isEqualTo(("==== Scenario: scenario icon:thumbs-down[role=\"red\",title=\"Failed\"]" + newLine() +
+                "description" + newLine() +
+                "" + newLine() +
+                "==========" + newLine() +
+                "Given::" + newLine() +
+
+                "passing step icon:thumbs-up[role=\"green\",title=\"Passed\"] [small right]#(000ms)#" + newLine() +
+
+                "When::" + newLine() +
+
+                "failing step icon:thumbs-down[role=\"red\",title=\"Failed\"] [small right]#(000ms)#" + newLine() +
+
+                "==========" + newLine() +
+                "" + newLine() +
+                "==== Scenario: scenario icon:thumbs-down[role=\"red\",title=\"Failed\"]" + newLine() +
+                "description" + newLine() +
+                "" + newLine() +
+                "==========" + newLine() +
+                "Then::" + newLine() +
+
+                "skipped step icon:thumbs-down[role=\"purple\",title=\"Skipped\"] [small right]#(000ms)#" + newLine() +
+
+                "==========" + newLine() +
+                "" + newLine() +
+                "==== Scenario: scenario icon:thumbs-down[role=\"red\",title=\"Failed\"]" + newLine() +
+                "description" + newLine() +
+                "" + newLine() +
+                "==========" + newLine() +
+                "Given::" + newLine() +
+
+                "undefined step icon:thumbs-down[role=\"yellow\",title=\"Undefined\"] [small right]#(000ms)#" + newLine() +
+
+                "==========" + newLine() +
+                "" + newLine()).replaceAll("\r", ""));
+    }
+
+    @Test
+    public void shouldRenderSourceDocStringInStep() {
+        List<Feature> features = FeatureParser.parse(featureWithSourceDocStringInStep);
+        final Feature feature = features.get(0);
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle(), new DocumentAttributes());
+        String resultDoc = featureRenderer.renderFeature(feature);
+        assertThat(resultDoc).isEqualTo(Expectations.FEATURE_WITH_SOURCE_DOC_STRING);
+    }
+
+    @Test
+    public void shouldNotGenerateSectionIdForFeatureBlankName() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithOneScenarioWithOnePassingStep();
+        feature.setName("    ");
+        List<Feature> features = new ArrayList<>();
+        features.add(feature);
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer();
+        assertThat(featureRenderer.renderFeatureSectionId(feature)).isEqualTo("");
+    }
+
+    @Test
+    public void shouldNotRenderSectionIdForFeatureWithNullName() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithOneScenarioWithOnePassingStep();
+        feature.setName(null);
+        List<Feature> features = new ArrayList<>();
+        features.add(feature);
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer();
+        assertThat(featureRenderer.renderFeatureSectionId(feature)).isEqualTo("");
+    }
+
+    @Test
+    public void shouldRenderSectionIdForFeatureWithNameWithSpaces() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithOneScenarioWithOnePassingStep();
+        feature.setName("Feature name");
+        List<Feature> features = new ArrayList<>();
+        features.add(feature);
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer();
+        assertThat(featureRenderer.renderFeatureSectionId(feature)).isEqualTo("[[Feature-name, Feature name]]");
+    }
+
+    @Test
+    public void shouldRenderSectionIdForFeatureWithNameWithSpacesAndComma() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithOneScenarioWithOnePassingStep();
+        feature.setName("Feature name, subname");
+        List<Feature> features = new ArrayList<>();
+        features.add(feature);
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer();
+        assertThat(featureRenderer.renderFeatureSectionId(feature)).isEqualTo("[[Feature-name-subname, Feature name, subname]]");
+    }
+
+    @Test
+    public void shouldRenderFeatureSectionId() {
+        final Feature feature = FeatureBuilder.instance().aFeatureWithOneScenarioWithOnePassingStep();
+        feature.setName("Name");
+        List<Feature> features = new ArrayList<>();
+        features.add(feature);
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer();
+        assertThat(featureRenderer.renderFeatureSectionId(feature)).isEqualTo("[[Name, Name]]");
+    }
+
+    @Test
+    public void shouldRenderFeatureDescription() {
+        final Feature feature = FeatureBuilder.instance().description("Feature description").name("Feature name").build();
+
+        CukedoctorFeatureRenderer featureRenderer = new CukedoctorFeatureRenderer(new CukedoctorDocumentBuilderImpl().nestTitle().nestTitle(), new DocumentAttributes());
+        featureRenderer = spy(featureRenderer);
+        doReturn("").when(featureRenderer).renderFeatureScenario(any(Scenario.class), eq(feature), any(CukedoctorDocumentBuilder.class));
+        String resultDoc = featureRenderer.renderFeature(feature);
+        assertThat(resultDoc).isEqualTo("[[Feature-name, Feature name]]" + newLine() +
+                "=== *Feature name*" + newLine() +
+                "" + newLine() +
+                "ifndef::backend-pdf[]" + newLine() +
+                "minmax::Feature-name[]" + newLine() +
+                "endif::[]" + newLine() +
+                "****" + newLine() +
+                "Feature description" + newLine() +
+                "****" + newLine() +
+                "" + newLine());
     }
 }
