+++ cukedoctor-converter/src/test/java/com/github/cukedoctor/renderer/Fixtures.java
@@ -7,11 +7,15 @@ public class Fixtures {
     public static final String embedDataDirectly;
     public static final String outline;
     public static final String invalidFeatureResult;
+    public static final String featureWithTableInStep;
+    public static final String featureWithSourceDocStringInStep;
 
     static {
         onePassingOneFailing = FileUtil.findJsonFile("target/test-classes/json-output/one_passing_one_failing.json");
         embedDataDirectly = FileUtil.findJsonFile("target/test-classes/json-output/embed_data_directly.json");
         outline = FileUtil.findJsonFile("target/test-classes/json-output/outline.json");
         invalidFeatureResult = FileUtil.findJsonFile("target/test-classes/json-output/invalid_feature_result.json");
+        featureWithTableInStep = FileUtil.findJsonFile("target/test-classes/json-output/step-with-table.json");
+        featureWithSourceDocStringInStep = FileUtil.findJsonFile("target/test-classes/json-output/step-with-source-doc-string.json");
     }
 }
